Refactoring API fetcher directory
