+++ commit-stalker/src/hooks/github-api/index.ts
@@ -1,0 +1,1 @@
+export const GITHUB_API_URL = 'https://api.github.com'
