+++ commit-stalker/src/pages/index.tsx
@@ -2,7 +2,7 @@ import Head from 'next/head'
 import { Readme } from '../components/organisms/readme'
 import { SearchForm } from '../components/organisms/search-form'
 import { Commits } from '../components/organisms/commits'
-import { useSearchQuery } from '../hooks/github-api'
+import { useSearchQuery } from '../hooks/github-api/use-search-query'
 import type { Commits as CommitsType } from '../models/commits'
 import { useRouter } from 'next/router'
 import { parseInputQuery } from '../lib/parse-input-query'
