+++ commit-stalker/tests/__mocks__/rest/create-handler.ts
@@ -1,4 +1,3 @@
-import { URL as GITHUB_API_URL } from '@/src/hooks/github-api'
 import { rest } from 'msw'
 
 type CreateHandlerMockParams = {
